CHORE: rename main/scrapper to main/instagram to be more accurately
