+++ src/main/instagram/scrapper.ts
@@ -1,4 +1,10 @@
-import puppeteer, { Browser, Page, TimeoutError } from 'puppeteer';
+import {
+  Browser,
+  ElementHandle,
+  Page,
+  Protocol,
+  TimeoutError,
+} from 'puppeteer';
 import { INSTAGRAM_URL } from './constants';
 import {
   DeactivatedIDError,
@@ -11,13 +17,8 @@ import { ScreenshotPath, URL } from './types';
 interface InsScarpper {
   login(userName: string, password: string): Promise<void>;
   exploreHashTag(tagName: string): Promise<Page>;
-  findPost(
-    page: Page,
-    postURL: URL
-  ): Promise<puppeteer.ElementHandle<HTMLAnchorElement>>;
-  makeRedBorder(
-    post: puppeteer.ElementHandle<HTMLAnchorElement>
-  ): Promise<void>;
+  findPost(page: Page, postURL: URL): Promise<ElementHandle<HTMLAnchorElement>>;
+  makeRedBorder(post: ElementHandle<HTMLAnchorElement>): Promise<void>;
   screenshot(page: Page, path: string): Promise<ScreenshotPath>;
   close(): Promise<void>;
 }
@@ -25,7 +26,7 @@ interface InsScarpper {
 class InsScarpperImpl implements InsScarpper {
   private browser: Browser;
   private URL = INSTAGRAM_URL;
-  private cookie: puppeteer.Protocol.Network.Cookie[] | null;
+  private cookie: Protocol.Network.Cookie[] | null;
 
   constructor(browser: Browser) {
     this.browser = browser;
@@ -82,7 +83,7 @@ class InsScarpperImpl implements InsScarpper {
   async findPost(
     page: Page,
     postURL: URL
-  ): Promise<puppeteer.ElementHandle<HTMLAnchorElement>> {
+  ): Promise<ElementHandle<HTMLAnchorElement>> {
     const postURLwithoutDomain = this.extractPostURL(postURL);
 
     const popularPostBox = await this.selectPopularPostBoxes(page);
@@ -96,7 +97,7 @@ class InsScarpperImpl implements InsScarpper {
     const post = allFindResult.filter(({ value }) => value !== null)[0].value;
 
     if (post !== undefined) {
-      return post as puppeteer.ElementHandle<HTMLAnchorElement>;
+      return post as ElementHandle<HTMLAnchorElement>;
     } else {
       throw new PostNotExistError(`포스트가 존재하지 않습니다: ${postURL}`);
     }
@@ -113,7 +114,7 @@ class InsScarpperImpl implements InsScarpper {
     }
   }
 
-  private selectHeader = async (page: puppeteer.Page) => {
+  private selectHeader = async (page: Page) => {
     const header = await page.$(`section > main > header`);
 
     if (header === null) {
@@ -128,7 +129,7 @@ class InsScarpperImpl implements InsScarpper {
   // 클라이언트 요구사항: 상위 9개 게시물 대상으로만 검색 및 스크린샷 촬영
   // 2023.09.02 기준 인스타그램 UI에서는 한 줄당 3개씩 총 28개 게시물이 노출
   // 따라서 상위 3개 줄만 추출하도록 함
-  private selectPopularPostBoxes = async (page: puppeteer.Page) => {
+  private selectPopularPostBoxes = async (page: Page) => {
     const allPopularPostBoxes = await page.$$(
       'section > main > article > div > div > div > div'
     );
@@ -174,7 +175,7 @@ class InsScarpperImpl implements InsScarpper {
     return screenshotPath;
   }
 
-  async makeRedBorder(post: puppeteer.ElementHandle<HTMLAnchorElement>) {
+  async makeRedBorder(post: ElementHandle<HTMLAnchorElement>) {
     await post.evaluate((post) => {
       post.style.display = 'block';
       post.style.outline = 'solid 5px red';
