+++ src/main/instagram/scrapperManager.ts
